+++ backend/src/main/java/reviewme/review/controller/ReviewApi.java
@@ -7,7 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import reviewme.review.dto.ReviewCreationResponse;
+import reviewme.review.dto.response.ReviewCreationResponse;
 import reviewme.review.dto.request.CreateReviewRequest;
 import reviewme.review.dto.response.ReviewDetailResponse;
