+++ backend/src/main/java/reviewme/review/controller/ReviewController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import reviewme.review.dto.ReviewCreationResponse;
+import reviewme.review.dto.response.ReviewCreationResponse;
 import reviewme.review.dto.request.CreateReviewRequest;
 import reviewme.review.dto.response.ReviewDetailResponse;
 import reviewme.review.service.ReviewService;
